requirements.txt
tvb-library~=2.0.3
plotly~=4.6.0
mne~=0.20.0
numpy~=1.18.2
matplotlib~=3.2.1
scipy~=1.4.1
pandas~=1.0.3

# Check version:  pip freeze | findstr plotly
# install tvb-data manually from https://zenodo.org/record/3688773#.Xqaa9ZnRYaE
